More complete support for address map configuration, only one format supported

+++ src/sst/elements/CramSim/c_BankCommand.hpp
@@ -48,7 +48,7 @@ class c_BankCommand {
 private:
 
 	unsigned m_seqNum;
-	unsigned m_addr;
+	ulong    m_addr;
 	unsigned m_row;
 	e_BankCommandType m_cmdMnemonic;
 	std::map<e_BankCommandType, std::string> m_cmdToString;
@@ -60,7 +60,7 @@ public:
 	//    friend std::ostream& operator<< (std::ostream& x_stream, const c_BankCommand& x_bankCommand);
 
 	explicit c_BankCommand(unsigned x_seqNum, e_BankCommandType x_cmdType,
-			unsigned x_addr);
+			ulong x_addr);
 
 	c_BankCommand(c_BankCommand&) = delete;
 	c_BankCommand(c_BankCommand&&) = delete;
