+++ src/sst/elements/CramSim/c_USimmTraceReader.hpp
@@ -65,7 +65,7 @@ private:
 	c_USimmTraceReader(const c_USimmTraceReader&); //do not implement
 	void operator=(const c_USimmTraceReader&);
 
-	c_Transaction* getNextTransaction(std::string x_txnType, unsigned x_addr, unsigned x_dataWidth);
+	c_Transaction* getNextTransaction(std::string x_txnType, ulong x_addr, unsigned x_dataWidth);
 
 	void createTxn();
